requirements.txt
Django>=3.1.0,<3.2.0
djangorestframework
psycopg2
Pillow

flake8